Move MESSAGE_QUEUE_PROVIDERS array to where it belongs - to msq_queue
The MESSAGE_QUEUE_PROVIDERS array was defined in the `__init__.py` of
the provider and it was pretty wrong because it was evaluated when
the provider was imported - so it was impossible to import the
base provider without initializing the arraym which caused circular
import.

Fixes: #51770
